+++ editor/plugins/particles_2d_editor_plugin.cpp
@@ -32,8 +32,8 @@
 
 #include "canvas_item_editor_plugin.h"
 #include "io/image_loader.h"
-#include "scene/3d/particles.h"
 #include "scene/gui/separator.h"
+#include "scene/resources/particles_material.h"
 
 void Particles2DEditorPlugin::edit(Object *p_object) {
