+++ editor/plugins/path_editor_plugin.cpp
@@ -30,7 +30,7 @@
 
 #include "path_editor_plugin.h"
 
-#include "os/keyboard.h"
+#include "core/os/keyboard.h"
 #include "scene/resources/curve.h"
 #include "spatial_editor_plugin.h"
